Made attenuator.
I'm not entirely sure what to call the inputs and outputs.  Which is the audio and which is the CV is completely arbitrary.  I could also make it work with an arbitrary number of signals, all multiplied together, but that may get needlessly complicated (eg if someone wants to make a GUI, a little hardwiring may be useful).