Was going to add cursor this way, but now I'm thinking better of it.
It'll be simpler to have separate patternNumber, eventRowNumber and channel variables.  Partly it would make looking in the patterns multidimensional array neater, using ints instead of arrays as keys of another array, and partly it would be more consistent once we have to include a similar cursor position for the current pattern in the pattern chain.