Tidying up of formulae.
Literal floats that happen to be integral numbers are now depicted as n.0 instead of float(n).

Some formulae are simplified by already working out the literal parts.

To convert sixtieths (eg numbers in the range 0 to +60) into volts in the range 0 to +5, they're divided by 60 and multiplied by 5.  To put it another way, they're divided by 12.  This seems kind of poetic to me, as there twelve semitones in an octave, so 1 / 12 == 5 / 60 == 0.08(3), which is how many more volts are required to go up a semitone in pitch.