Files are now simpler.
None of this shifting by 48 malarkey.  If you want to debug the files, use a hex viewer.  This way, we're not limited by a range of 0 to 60 for absolutely *everything*.  So we can now have, for instance, pattern lengths of up to 255 rows, implementation permitting.  We could even store eight channels of gates, for percussion, using a single byte per row.