Factory reset related bugfix.
The factory reset adding commit
(1dae338314fcb4bc602ca7b0ad012920ba9561e8) introduced a bug which
caused Stepper Acid to constantly output its status over the serial
interface.  This fixes that issue.

Now that the input variable is a signed int, it is large enough to
spot any genuine input in the range of 0 to 255, and on top of that
also accept -2 and -1 as examples of things that can't possibly exist
as real input bytes, the same way the EOF signal in <stdio.h> works.